@@ -6,6 +6,7 @@ import static com.yugabyte.yw.common.ModelFactory.newProvider;
 import static com.yugabyte.yw.common.TestHelper.createTempFile;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsInAnyOrder;
+import static org.hamcrest.Matchers.either;
 import static org.hamcrest.Matchers.emptyString;
 import static org.hamcrest.Matchers.hasEntry;
 import static org.hamcrest.Matchers.is;
@@ -19,23 +20,24 @@ import com.yugabyte.yba.v2.client.Configuration;
 import com.yugabyte.yba.v2.client.api.UniverseManagementApi;
 import com.yugabyte.yba.v2.client.models.AuditLogConfig;
 import com.yugabyte.yba.v2.client.models.AvailabilityZoneGFlags;
+import com.yugabyte.yba.v2.client.models.CloudSpecificInfo;
 import com.yugabyte.yba.v2.client.models.ClusterGFlags;
 import com.yugabyte.yba.v2.client.models.ClusterInfo;
 import com.yugabyte.yba.v2.client.models.ClusterNetworkingSpec;
 import com.yugabyte.yba.v2.client.models.ClusterNetworkingSpec.EnableExposingServiceEnum;
+import com.yugabyte.yba.v2.client.models.ClusterPlacementSpec;
 import com.yugabyte.yba.v2.client.models.ClusterProviderSpec;
 import com.yugabyte.yba.v2.client.models.ClusterSpec;
 import com.yugabyte.yba.v2.client.models.ClusterSpec.ClusterTypeEnum;
 import com.yugabyte.yba.v2.client.models.ClusterStorageSpec;
 import com.yugabyte.yba.v2.client.models.ClusterStorageSpec.StorageTypeEnum;
 import com.yugabyte.yba.v2.client.models.CommunicationPortsSpec;
+import com.yugabyte.yba.v2.client.models.EncryptionAtRestInfo;
 import com.yugabyte.yba.v2.client.models.EncryptionAtRestSpec;
-import com.yugabyte.yba.v2.client.models.EncryptionAtRestSpec.OpTypeEnum;
-import com.yugabyte.yba.v2.client.models.EncryptionAtRestSpec.TypeEnum;
 import com.yugabyte.yba.v2.client.models.EncryptionInTransitSpec;
+import com.yugabyte.yba.v2.client.models.NodeDetails;
 import com.yugabyte.yba.v2.client.models.PlacementAZ;
 import com.yugabyte.yba.v2.client.models.PlacementCloud;
-import com.yugabyte.yba.v2.client.models.PlacementInfo;
 import com.yugabyte.yba.v2.client.models.PlacementRegion;
 import com.yugabyte.yba.v2.client.models.UniverseCreateSpec;
 import com.yugabyte.yba.v2.client.models.UniverseCreateSpec.ArchEnum;
@@ -43,8 +45,10 @@ import com.yugabyte.yba.v2.client.models.UniverseCreateSpecYcql;
 import com.yugabyte.yba.v2.client.models.UniverseCreateSpecYsql;
 import com.yugabyte.yba.v2.client.models.UniverseInfo;
 import com.yugabyte.yba.v2.client.models.UniverseLogsExporterConfig;
+import com.yugabyte.yba.v2.client.models.UniverseNetworkingSpec;
 import com.yugabyte.yba.v2.client.models.UniverseResp;
 import com.yugabyte.yba.v2.client.models.UniverseSpec;
+import com.yugabyte.yba.v2.client.models.XClusterInfo;
 import com.yugabyte.yba.v2.client.models.YBPTask;
 import com.yugabyte.yba.v2.client.models.YCQLAuditConfig;
 import com.yugabyte.yba.v2.client.models.YSQLAuditConfig;
@@ -59,10 +63,12 @@ import com.yugabyte.yw.common.FakeDBApplication;
 import com.yugabyte.yw.common.TestHelper;
 import com.yugabyte.yw.common.certmgmt.CertConfigType;
 import com.yugabyte.yw.common.gflags.SpecificGFlags;
+import com.yugabyte.yw.common.kms.util.EncryptionAtRestUtil;
 import com.yugabyte.yw.controllers.UniverseControllerTestBase;
 import com.yugabyte.yw.forms.EncryptionAtRestConfig;
 import com.yugabyte.yw.forms.UniverseDefinitionTaskParams;
 import com.yugabyte.yw.forms.UniverseDefinitionTaskParams.Cluster;
+import com.yugabyte.yw.forms.UniverseDefinitionTaskParams.UserIntent;
 import com.yugabyte.yw.forms.UniverseTaskParams.CommunicationPorts;
 import com.yugabyte.yw.models.AvailabilityZone;
 import com.yugabyte.yw.models.CertificateInfo;
@@ -74,10 +80,14 @@ import com.yugabyte.yw.models.Universe;
 import com.yugabyte.yw.models.helpers.TaskType;
 import java.io.IOException;
 import java.security.NoSuchAlgorithmException;
+import java.time.OffsetDateTime;
+import java.time.ZoneOffset;
+import java.time.temporal.ChronoUnit;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -190,7 +200,7 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
   }
 
   @Test
-  public void testGetV2Universe() throws ApiException {
+  public void testGetUniverseV2() throws ApiException {
     UUID uUUID = createUniverse(customer.getId()).getUniverseUUID();
     Universe dbUniverse =
         Universe.saveDetails(
@@ -263,6 +273,82 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
     return auditLogConfig;
   }
 
+  private PlacementRegion getOrCreatePlacementRegion(PlacementCloud placementCloud, Region region) {
+    PlacementRegion placementRegion = null;
+    if (placementCloud.getRegionList() == null) {
+      placementCloud.setRegionList(new ArrayList<>());
+    }
+    Optional<PlacementRegion> optPlacementRegion =
+        placementCloud.getRegionList().stream()
+            .filter(r -> r.getUuid().equals(region.getUuid()))
+            .findAny();
+    if (optPlacementRegion.isEmpty()) {
+      placementRegion =
+          new PlacementRegion()
+              .code(region.getCode())
+              .name(region.getName())
+              .uuid(region.getUuid());
+      placementCloud.addRegionListItem(placementRegion);
+    } else {
+      placementRegion = optPlacementRegion.get();
+    }
+    return placementRegion;
+  }
+
+  private PlacementAZ getOrCreatePlacementAz(
+      PlacementRegion placementRegion, AvailabilityZone zone) {
+    PlacementAZ placementAZ = null;
+    if (placementRegion.getAzList() == null) {
+      placementRegion.setAzList(new ArrayList<>());
+    }
+    Optional<PlacementAZ> optPlacementAz =
+        placementRegion.getAzList().stream()
+            .filter(z -> z.getUuid().equals(zone.getUuid()))
+            .findAny();
+    if (optPlacementAz.isEmpty()) {
+      placementAZ =
+          new PlacementAZ()
+              .uuid(zone.getUuid())
+              .name(zone.getName())
+              .subnet(zone.getSubnet())
+              .secondarySubnet(zone.getSecondarySubnet())
+              .leaderAffinity(true);
+      placementRegion.addAzListItem(placementAZ);
+    } else {
+      placementAZ = optPlacementAz.get();
+    }
+    return placementAZ;
+  }
+
+  private PlacementCloud fromProvider(int numNodesToPlace, int rfTotal) {
+    Provider provider = Provider.get(customer.getUuid(), providerUuid);
+    PlacementCloud placementCloud =
+        new PlacementCloud().uuid(providerUuid).code(provider.getCode());
+    Region region = Region.getByProvider(providerUuid).get(0);
+    PlacementRegion pr = getOrCreatePlacementRegion(placementCloud, region);
+    int numPlacedNodes = 0;
+    int rf = 0;
+    while (numPlacedNodes < numNodesToPlace) {
+      List<AvailabilityZone> zones = AvailabilityZone.getAZsForRegion(region.getUuid());
+      for (AvailabilityZone zone : zones) {
+        PlacementAZ pAz = getOrCreatePlacementAz(pr, zone);
+        int numNodesInAz = pAz.getNumNodesInAz() != null ? pAz.getNumNodesInAz() : 0;
+        pAz.numNodesInAz(numNodesInAz + 1);
+        if (rf < rfTotal) {
+          // increment rf in this az
+          int currRf = pAz.getReplicationFactor() != null ? pAz.getReplicationFactor() : 0;
+          pAz.setReplicationFactor(currRf + 1);
+          rf++;
+        }
+        numPlacedNodes++;
+        if (numPlacedNodes >= numNodesToPlace) {
+          break;
+        }
+      }
+    }
+    return placementCloud;
+  }
+
   @Test
   public void testCreateUniverseV2() throws ApiException {
     UniverseManagementApi api = new UniverseManagementApi();
@@ -273,12 +359,19 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
     UniverseSpec universeSpec = new UniverseSpec();
     universeCreateSpec.spec(universeSpec);
     universeSpec.name("Test-V2-Universe");
+    universeSpec.setYbSoftwareVersion("2.20.0.0-b123");
+    universeSpec.setUseTimeSync(true);
+    universeSpec.setUseSystemd(true);
     universeSpec.ysql(new YsqlSpec().enable(true).enableAuth(false));
     universeSpec.ycql(new YcqlSpec().enable(true).enableAuth(false));
-    universeSpec.communicationPorts(
-        new CommunicationPortsSpec().masterHttpPort(1234).tserverHttpPort(5678));
-    EncryptionAtRestSpec ear = new EncryptionAtRestSpec();
-    ear.kmsConfigUuid(UUID.randomUUID()).opType(OpTypeEnum.ENABLE).type(TypeEnum.CMK);
+    universeSpec.networkingSpec(
+        new UniverseNetworkingSpec()
+            .assignPublicIp(true)
+            .assignStaticPublicIp(true)
+            .enableIpv6(false)
+            .communicationPorts(
+                new CommunicationPortsSpec().masterHttpPort(1234).tserverHttpPort(5678)));
+    EncryptionAtRestSpec ear = new EncryptionAtRestSpec().kmsConfigUuid(UUID.randomUUID());
     universeSpec.encryptionAtRestSpec(ear);
     EncryptionInTransitSpec eit = new EncryptionInTransitSpec();
     eit.enableNodeToNodeEncrypt(true)
@@ -291,11 +384,12 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
     primaryClusterSpec.setNumNodes(6);
     primaryClusterSpec.setInstanceType(ApiUtils.UTIL_INST_TYPE);
     primaryClusterSpec.setReplicationFactor(5);
-    primaryClusterSpec.setYbSoftwareVersion("2.20.0.0-b123");
-    primaryClusterSpec.setUseTimeSync(true);
     primaryClusterSpec.setUseSpotInstance(true);
-    primaryClusterSpec.setUseSystemd(true);
     primaryClusterSpec.setInstanceTags(Map.of("itag1", "ival1", "itag2", "ival2"));
+    PlacementCloud placementCloud =
+        fromProvider(primaryClusterSpec.getNumNodes(), primaryClusterSpec.getReplicationFactor());
+    primaryClusterSpec.setPlacementSpec(
+        new ClusterPlacementSpec().cloudList(List.of(placementCloud)));
     ClusterProviderSpec providerSpec = new ClusterProviderSpec();
     providerSpec.setProvider(providerUuid);
     providerSpec.setRegionList(List.of(Region.getByProvider(providerUuid).get(0).getUuid()));
@@ -306,9 +400,6 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
         new ClusterStorageSpec().volumeSize(54321).numVolumes(2).storageType(StorageTypeEnum.GP2));
     primaryClusterSpec.setNetworkingSpec(
         new ClusterNetworkingSpec()
-            .assignPublicIp(true)
-            .assignStaticPublicIp(true)
-            .enableIpv6(false)
             .enableLb(true)
             .enableExposingService(EnableExposingServiceEnum.EXPOSED));
     primaryClusterSpec.setGflags(createPrimaryClusterGFlags());
@@ -317,12 +408,18 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
     UUID fakeTaskUUID = FakeDBApplication.buildTaskInfo(null, TaskType.CreateUniverse);
     when(mockCommissioner.submit(any(TaskType.class), any(UniverseDefinitionTaskParams.class)))
         .thenReturn(fakeTaskUUID);
+    when(mockRuntimeConfig.getInt("yb.universe.otel_collector_metrics_port")).thenReturn(8889);
     YBPTask createTask = api.createUniverse(customer.getUuid(), universeCreateSpec);
     UUID universeUUID = createTask.getResourceUuid();
-    Universe dbUniv = Universe.getOrBadRequest(universeUUID, customer);
+    Universe dbUniverse = Universe.getOrBadRequest(universeUUID, customer);
 
     // validate that the Universe created in the DB matches properties specified in the createSpec
-    validateUniverseCreateSpec(universeCreateSpec, dbUniv);
+    validateUniverseCreateSpec(universeCreateSpec, dbUniverse);
+
+    // fetch the newly created Universe using v2 API and validate it once more
+    UniverseResp universeResp = api.getUniverse(customer.getUuid(), universeUUID);
+    validateUniverseSpec(universeResp.getSpec(), dbUniverse);
+    validateUniverseInfo(universeResp.getInfo(), dbUniverse);
   }
 
   private void validateUniverseCreateSpec(UniverseCreateSpec v2Univ, Universe dbUniv) {
@@ -332,7 +429,19 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
   }
 
   private void validateUniverseSpec(UniverseSpec v2UnivSpec, Universe dbUniv) {
+    UserIntent dbUserIntent = dbUniv.getUniverseDetails().getPrimaryCluster().userIntent;
     assertThat(v2UnivSpec.getName(), is(dbUniv.getName()));
+    assertThat(v2UnivSpec.getYbSoftwareVersion(), is(dbUserIntent.ybSoftwareVersion));
+    if (v2UnivSpec.getUseSystemd() == null) {
+      assertThat(dbUserIntent.useSystemd, is(false));
+    } else {
+      assertThat(v2UnivSpec.getUseSystemd(), is(dbUserIntent.useSystemd));
+    }
+    if (v2UnivSpec.getUseTimeSync() == null) {
+      assertThat(dbUserIntent.useTimeSync, is(false));
+    } else {
+      assertThat(v2UnivSpec.getUseTimeSync(), is(dbUserIntent.useTimeSync));
+    }
     if (v2UnivSpec.getOverridePrebuiltAmiDbVersion() == null) {
       assertThat(dbUniv.getUniverseDetails().overridePrebuiltAmiDBVersion, is(false));
     } else {
@@ -346,8 +455,7 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
       assertThat(
           v2UnivSpec.getRemotePackagePath(), is(dbUniv.getUniverseDetails().remotePackagePath));
     }
-    validateCommunicationPorts(
-        v2UnivSpec.getCommunicationPorts(), dbUniv.getUniverseDetails().communicationPorts);
+    validateUniverseNetworkginSpec(v2UnivSpec.getNetworkingSpec(), dbUniv.getUniverseDetails());
     validateEncryptionAtRest(
         v2UnivSpec.getEncryptionAtRestSpec(), dbUniv.getUniverseDetails().encryptionAtRestConfig);
     validateEncryptionInTransit(
@@ -355,6 +463,28 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
     validateYsqlSpec(v2UnivSpec.getYsql(), dbUniv.getUniverseDetails());
     validateYcqlSpec(v2UnivSpec.getYcql(), dbUniv.getUniverseDetails());
     validateClusters(v2UnivSpec.getClusters(), dbUniv.getUniverseDetails().clusters);
+  }
+
+  private void validateUniverseNetworkginSpec(
+      UniverseNetworkingSpec v2NetworkingSpec, UniverseDefinitionTaskParams dbUniv) {
+    UserIntent primaryUserIntent = dbUniv.getPrimaryCluster().userIntent;
+    if (v2NetworkingSpec.getAssignPublicIp() == null) {
+      assertThat(primaryUserIntent.assignPublicIP, is(true));
+    } else {
+      assertThat(v2NetworkingSpec.getAssignPublicIp(), is(primaryUserIntent.assignPublicIP));
+    }
+    if (v2NetworkingSpec.getAssignStaticPublicIp() == null) {
+      assertThat(primaryUserIntent.assignStaticPublicIP, is(false));
+    } else {
+      assertThat(
+          v2NetworkingSpec.getAssignStaticPublicIp(), is(primaryUserIntent.assignStaticPublicIP));
+    }
+    if (v2NetworkingSpec.getEnableIpv6() == null) {
+      assertThat(primaryUserIntent.enableIPV6, is(false));
+    } else {
+      assertThat(v2NetworkingSpec.getEnableIpv6(), is(primaryUserIntent.enableIPV6));
+    }
+    validateCommunicationPorts(v2NetworkingSpec.getCommunicationPorts(), dbUniv.communicationPorts);
   }
 
   private void validateCommunicationPorts(CommunicationPortsSpec v2CP, CommunicationPorts dbCP) {
@@ -433,9 +563,7 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
 
   private void validateEncryptionAtRest(EncryptionAtRestSpec v2Enc, EncryptionAtRestConfig dbEnc) {
     assertThat(v2Enc.getKmsConfigUuid(), is(dbEnc.kmsConfigUUID));
-    assertThat(v2Enc.getOpType().getValue(), is(dbEnc.opType.name()));
-    assertThat(v2Enc.getType().getValue(), is(dbEnc.type.name()));
-    assertThat(dbEnc.encryptionAtRestEnabled, is(v2Enc.getOpType().equals(OpTypeEnum.ENABLE)));
+    assertThat(dbEnc.type, is(EncryptionAtRestUtil.KeyType.DATA_KEY));
   }
 
   private void validateEncryptionInTransit(
@@ -489,10 +617,8 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
     assertThat(v2Cluster.getReplicationFactor(), is(dbCluster.userIntent.replicationFactor));
     assertThat(v2Cluster.getInstanceType(), is(dbCluster.userIntent.instanceType));
     assertThat(v2Cluster.getUseSpotInstance(), is(dbCluster.userIntent.useSpotInstance));
-    assertThat(v2Cluster.getUseSystemd(), is(dbCluster.userIntent.useSystemd));
-    assertThat(v2Cluster.getUseTimeSync(), is(dbCluster.userIntent.useTimeSync));
-    assertThat(v2Cluster.getYbSoftwareVersion(), is(dbCluster.userIntent.ybSoftwareVersion));
     validateProviderSpec(v2Cluster.getProviderSpec(), dbCluster);
+    validatePlacementSpec(v2Cluster.getPlacementSpec(), dbCluster.placementInfo);
     validateStorageSpec(v2Cluster.getStorageSpec(), dbCluster);
     validateNetworkingSpec(v2Cluster.getNetworkingSpec(), dbCluster);
     validateGFlags(v2Cluster.getGflags(), dbCluster.userIntent.specificGFlags);
@@ -508,6 +634,30 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
     assertThat(
         v2ProviderSpec.getRegionList(),
         containsInAnyOrder(dbCluster.userIntent.regionList.toArray()));
+  }
+
+  private void validatePlacementSpec(
+      ClusterPlacementSpec v2PI, com.yugabyte.yw.models.helpers.PlacementInfo dbPI) {
+    if (v2PI == null) {
+      assertThat(dbPI, is(nullValue()));
+      return;
+    }
+    if (v2PI.getCloudList() == null) {
+      assertThat(dbPI.cloudList, is(nullValue()));
+      return;
+    }
+    assertThat(v2PI.getCloudList().size(), is(dbPI.cloudList.size()));
+    for (PlacementCloud v2Cloud : v2PI.getCloudList()) {
+      // find corresponding cloud in db cloud list
+      com.yugabyte.yw.models.helpers.PlacementInfo.PlacementCloud dbCloud =
+          dbPI.cloudList.stream()
+              .filter(c -> c.uuid.equals(v2Cloud.getUuid()))
+              .findFirst()
+              .orElseThrow();
+      assertThat(v2Cloud.getCode(), is(dbCloud.code));
+      assertThat(v2Cloud.getDefaultRegion(), is(dbCloud.defaultRegion));
+      verifyPlacementRegion(v2Cloud.getRegionList(), dbCloud.regionList);
+    }
   }
 
   private void validateStorageSpec(ClusterStorageSpec v2StorageSpec, Cluster dbCluster) {
@@ -527,17 +677,9 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
   }
 
   private void validateNetworkingSpec(ClusterNetworkingSpec v2NetworkingSpec, Cluster dbCluster) {
-    assertThat(v2NetworkingSpec.getAssignPublicIp(), is(dbCluster.userIntent.assignPublicIP));
-    assertThat(
-        v2NetworkingSpec.getAssignStaticPublicIp(), is(dbCluster.userIntent.assignStaticPublicIP));
     assertThat(
         v2NetworkingSpec.getEnableExposingService().getValue(),
         is(dbCluster.userIntent.enableExposingService.name()));
-    if (v2NetworkingSpec.getEnableIpv6() == null) {
-      assertThat(dbCluster.userIntent.enableIPV6, is(false));
-    } else {
-      assertThat(v2NetworkingSpec.getEnableIpv6(), is(dbCluster.userIntent.enableIPV6));
-    }
     assertThat(v2NetworkingSpec.getEnableLb(), is(dbCluster.userIntent.enableLB));
   }
 
@@ -748,6 +890,45 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
       assertThat(v2UnivInfo.getArch().getValue(), is(dbUniv.arch.name()));
     }
     validateClustersInfo(v2UnivInfo.getClusters(), dbUniverse);
+    OffsetDateTime dbCreationDate =
+        dbUniverse
+            .getCreationDate()
+            .toInstant()
+            .truncatedTo(ChronoUnit.SECONDS)
+            .atOffset(ZoneOffset.UTC);
+    assertThat(v2UnivInfo.getCreationDate().compareTo(dbCreationDate), is(0));
+    com.yugabyte.yw.forms.UniverseResp v1UniverseResp =
+        new com.yugabyte.yw.forms.UniverseResp(dbUniverse);
+    assertThat(
+        v2UnivInfo.getDnsName(),
+        is(v1UniverseResp.getDnsName(customer, Provider.getOrBadRequest(providerUuid))));
+    validateEncryptionAtRestInfo(
+        v2UnivInfo.getEncryptionAtRestInfo(), dbUniv.encryptionAtRestConfig);
+    assertThat(
+        v2UnivInfo.getIsKubernetesOperatorControlled(), is(dbUniv.isKubernetesOperatorControlled));
+    assertThat(v2UnivInfo.getIsSoftwareRollbackAllowed(), is(dbUniv.isSoftwareRollbackAllowed));
+    validateNodeDetailsSet(v2UnivInfo.getNodeDetailsSet(), dbUniv.nodeDetailsSet);
+    assertThat(v2UnivInfo.getNodePrefix(), is(dbUniv.nodePrefix));
+    assertThat(v2UnivInfo.getNodesResizeAvailable(), is(dbUniv.nodesResizeAvailable));
+    assertThat(v2UnivInfo.getOtelCollectorEnabled(), is(dbUniv.otelCollectorEnabled));
+    assertThat(
+        v2UnivInfo.getPlacementModificationTaskUuid(), is(dbUniv.placementModificationTaskUuid));
+    assertThat(v2UnivInfo.getPlatformUrl(), is(dbUniv.platformUrl));
+    assertThat(
+        v2UnivInfo.getSoftwareUpgradeState().getValue(), is(dbUniv.softwareUpgradeState.name()));
+    assertThat(v2UnivInfo.getUniversePaused(), is(dbUniv.universePaused));
+    assertThat(v2UnivInfo.getUniverseUuid(), is(dbUniverse.getUniverseUUID()));
+    assertThat(v2UnivInfo.getUpdateInProgress(), is(dbUniv.updateInProgress));
+    assertThat(v2UnivInfo.getUpdateSucceeded(), is(dbUniv.updateSucceeded));
+    if (v2UnivInfo.getUpdatingTask() == null) {
+      assertThat(dbUniv.updatingTask, is(nullValue()));
+    } else {
+      assertThat(v2UnivInfo.getUpdatingTask(), is(dbUniv.updatingTask.name()));
+    }
+    assertThat(v2UnivInfo.getUpdatingTaskUuid(), is(dbUniv.updatingTaskUUID));
+    assertThat(v2UnivInfo.getVersion(), is(dbUniverse.getVersion()));
+    assertThat(v2UnivInfo.getYbcSoftwareVersion(), is(dbUniv.getYbcSoftwareVersion()));
+    validateXClusterInfo(v2UnivInfo.getxClusterInfo(), dbUniv.xClusterInfo);
   }
 
   private void validateClustersInfo(List<ClusterInfo> v2Clusters, Universe dbUniverse) {
@@ -767,30 +948,17 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
   private void validateClusterInfo(ClusterInfo v2Cluster, Cluster dbCluster) {
     assertThat(v2Cluster.getUuid(), is(dbCluster.uuid));
     assertThat(v2Cluster.getSpotPrice(), is(dbCluster.userIntent.spotPrice));
-    validatePlacementInfo(v2Cluster.getPlacementInfo(), dbCluster.placementInfo);
   }
 
-  private void validatePlacementInfo(
-      PlacementInfo v2PI, com.yugabyte.yw.models.helpers.PlacementInfo dbPI) {
-    if (v2PI == null) {
-      assertThat(dbPI, is(nullValue()));
-      return;
-    }
-    if (v2PI.getCloudList() == null) {
-      assertThat(dbPI.cloudList, is(nullValue()));
-      return;
-    }
-    assertThat(v2PI.getCloudList().size(), is(dbPI.cloudList.size()));
-    for (PlacementCloud v2Cloud : v2PI.getCloudList()) {
-      // find corresponding cloud in db cloud list
-      com.yugabyte.yw.models.helpers.PlacementInfo.PlacementCloud dbCloud =
-          dbPI.cloudList.stream()
-              .filter(c -> c.uuid.equals(v2Cloud.getUuid()))
-              .findFirst()
-              .orElseThrow();
-      assertThat(v2Cloud.getCode(), is(dbCloud.code));
-      assertThat(v2Cloud.getDefaultRegion(), is(dbCloud.defaultRegion));
-      verifyPlacementRegion(v2Cloud.getRegionList(), dbCloud.regionList);
+  private void validateEncryptionAtRestInfo(
+      EncryptionAtRestInfo v2EncryptionAtRestInfo,
+      EncryptionAtRestConfig dbEncryptionAtRestConfig) {
+    if (v2EncryptionAtRestInfo.getEncryptionAtRestStatus() == null) {
+      assertThat(dbEncryptionAtRestConfig.encryptionAtRestEnabled, is(false));
+    } else {
+      assertThat(
+          v2EncryptionAtRestInfo.getEncryptionAtRestStatus(),
+          is(dbEncryptionAtRestConfig.encryptionAtRestEnabled));
     }
   }
 
@@ -826,7 +994,7 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
       // find corresponding az in db az list
       com.yugabyte.yw.models.helpers.PlacementInfo.PlacementAZ dbAz =
           dbAzList.stream().filter(a -> a.uuid.equals(v2Az.getUuid())).findFirst().orElseThrow();
-      assertThat(v2Az.getIsAffinitized(), is(dbAz.isAffinitized));
+      assertThat(v2Az.getLeaderAffinity(), is(dbAz.isAffinitized));
       assertThat(v2Az.getLbName(), is(dbAz.lbName));
       assertThat(v2Az.getName(), is(dbAz.name));
       assertThat(v2Az.getNumNodesInAz(), is(dbAz.numNodesInAZ));
@@ -834,5 +1002,106 @@ public class UniverseManagementApiControllerImpTest extends UniverseControllerTe
       assertThat(v2Az.getSecondarySubnet(), is(dbAz.secondarySubnet));
       assertThat(v2Az.getSubnet(), is(dbAz.subnet));
     }
+  }
+
+  private void validateNodeDetailsSet(
+      List<NodeDetails> v2NodeDetailsSet,
+      Set<com.yugabyte.yw.models.helpers.NodeDetails> dbNodeDetailsSet) {
+    if (v2NodeDetailsSet == null || v2NodeDetailsSet.size() == 0) {
+      assertThat(dbNodeDetailsSet, is(either(nullValue()).or(is(Set.of()))));
+      return;
+    }
+    for (NodeDetails v2NodeDetail : v2NodeDetailsSet) {
+      // find the corresponding NodeDetail in dbNodeDetailSet
+      com.yugabyte.yw.models.helpers.NodeDetails dbNodeDetail =
+          dbNodeDetailsSet.stream()
+              .filter(n -> v2NodeDetail.getNodeIdx().equals(n.nodeIdx))
+              .findAny()
+              .orElseThrow();
+      validateNodeDetail(v2NodeDetail, dbNodeDetail);
+    }
+  }
+
+  private void validateNodeDetail(
+      NodeDetails v2NodeDetail, com.yugabyte.yw.models.helpers.NodeDetails dbNodeDetail) {
+    if (v2NodeDetail == null) {
+      assertThat(dbNodeDetail, is(nullValue()));
+      return;
+    }
+    assertThat(v2NodeDetail.getAzUuid(), is(dbNodeDetail.azUuid));
+    validateCloudSpecificInfo(v2NodeDetail.getCloudInfo(), dbNodeDetail.cloudInfo);
+    assertThat(v2NodeDetail.getCronsActive(), is(dbNodeDetail.cronsActive));
+    if (v2NodeDetail.getDedicatedTo() == null) {
+      assertThat(dbNodeDetail.dedicatedTo, is(nullValue()));
+    } else {
+      assertThat(v2NodeDetail.getDedicatedTo().getValue(), is(dbNodeDetail.dedicatedTo.name()));
+    }
+    assertThat(v2NodeDetail.getDisksAreMountedByUuid(), is(dbNodeDetail.disksAreMountedByUUID));
+    assertThat(v2NodeDetail.getIsMaster(), is(dbNodeDetail.isMaster));
+    assertThat(v2NodeDetail.getIsRedisServer(), is(dbNodeDetail.isRedisServer));
+    assertThat(v2NodeDetail.getIsTserver(), is(dbNodeDetail.isTserver));
+    assertThat(v2NodeDetail.getIsYqlServer(), is(dbNodeDetail.isYqlServer));
+    assertThat(v2NodeDetail.getIsYsqlServer(), is(dbNodeDetail.isYsqlServer));
+    if (v2NodeDetail.getLastVolumeUpdateTime() == null) {
+      assertThat(dbNodeDetail.lastVolumeUpdateTime, is(nullValue()));
+    } else {
+      // might need to use compareTo() method instead?
+      assertThat(v2NodeDetail.getLastVolumeUpdateTime(), is(dbNodeDetail.lastVolumeUpdateTime));
+    }
+    assertThat(v2NodeDetail.getMachineImage(), is(dbNodeDetail.machineImage));
+    if (v2NodeDetail.getMasterState() == null) {
+      assertThat(dbNodeDetail.masterState, is(nullValue()));
+    } else {
+      assertThat(v2NodeDetail.getMasterState().getValue(), is(dbNodeDetail.masterState.name()));
+    }
+    assertThat(v2NodeDetail.getNodeIdx(), is(dbNodeDetail.nodeIdx));
+    assertThat(v2NodeDetail.getNodeName(), is(dbNodeDetail.nodeName));
+    assertThat(v2NodeDetail.getPlacementUuid(), is(dbNodeDetail.placementUuid));
+    assertThat(v2NodeDetail.getSshPortOverride(), is(dbNodeDetail.sshPortOverride));
+    assertThat(v2NodeDetail.getSshUserOverride(), is(dbNodeDetail.sshUserOverride));
+    if (v2NodeDetail.getState() == null) {
+      assertThat(dbNodeDetail.state, is(nullValue()));
+    } else {
+      assertThat(v2NodeDetail.getState().getValue(), is(dbNodeDetail.state.name()));
+    }
+    assertThat(v2NodeDetail.getYbPrebuiltAmi(), is(dbNodeDetail.ybPrebuiltAmi));
+  }
+
+  private void validateCloudSpecificInfo(
+      CloudSpecificInfo v2CloudInfo, com.yugabyte.yw.models.helpers.CloudSpecificInfo dbCloudInfo) {
+    if (v2CloudInfo == null) {
+      assertThat(dbCloudInfo, is(nullValue()));
+      return;
+    }
+    if (v2CloudInfo.getAssignPublicIp() == null) {
+      assertThat(dbCloudInfo.assignPublicIP, is(true));
+    } else {
+      assertThat(v2CloudInfo.getAssignPublicIp(), is(dbCloudInfo.assignPublicIP));
+    }
+    assertThat(v2CloudInfo.getAz(), is(dbCloudInfo.az));
+    assertThat(v2CloudInfo.getCloud(), is(dbCloudInfo.cloud));
+    assertThat(v2CloudInfo.getInstanceType(), is(dbCloudInfo.instance_type));
+    assertThat(v2CloudInfo.getKubernetesNamespace(), is(dbCloudInfo.kubernetesNamespace));
+    assertThat(v2CloudInfo.getKubernetesPodName(), is(dbCloudInfo.kubernetesPodName));
+    if (v2CloudInfo.getLunIndexes() != null) {
+      assertThat(v2CloudInfo.getLunIndexes(), containsInAnyOrder(dbCloudInfo.lun_indexes));
+    }
+    assertThat(v2CloudInfo.getMountRoots(), is(dbCloudInfo.mount_roots));
+    assertThat(v2CloudInfo.getPrivateDns(), is(dbCloudInfo.private_dns));
+    assertThat(v2CloudInfo.getPrivateIp(), is(dbCloudInfo.private_ip));
+    assertThat(v2CloudInfo.getPublicDns(), is(dbCloudInfo.public_dns));
+    assertThat(v2CloudInfo.getPublicIp(), is(dbCloudInfo.public_ip));
+    assertThat(v2CloudInfo.getRegion(), is(dbCloudInfo.region));
+    assertThat(v2CloudInfo.getRootVolume(), is(dbCloudInfo.root_volume));
+    assertThat(v2CloudInfo.getSecondaryPrivateIp(), is(dbCloudInfo.secondary_private_ip));
+    assertThat(v2CloudInfo.getSecondarySubnetId(), is(dbCloudInfo.secondary_subnet_id));
+    assertThat(v2CloudInfo.getSubnetId(), is(dbCloudInfo.subnet_id));
+    assertThat(v2CloudInfo.getUseTimeSync(), is(dbCloudInfo.useTimeSync));
+  }
+
+  private void validateXClusterInfo(
+      XClusterInfo getxClusterInfo,
+      com.yugabyte.yw.forms.UniverseDefinitionTaskParams.XClusterInfo xClusterInfo) {
+    // TODO Auto-generated method stub
   }
 }
